+++ packages/util-nonempty/rollup.config.js
@@ -1,0 +1,38 @@
+import resolve from 'rollup-plugin-node-resolve'
+import commonjs from 'rollup-plugin-commonjs'
+import babel from 'rollup-plugin-babel'
+import pkg from './package.json'
+
+export default [
+  // browser-friendly UMD build
+  {
+    input: 'src/main.js',
+    output: {
+      file: pkg.browser,
+      format: 'umd',
+      name: 'utilNonempty',
+    },
+    plugins: [
+      resolve(), // so Rollup can find deps
+      commonjs(), // so Rollup can convert deps to an ES modules
+      babel({
+        exclude: ['node_modules/**'],
+      }),
+    ],
+  },
+
+  // CommonJS (for Node) and ES module (for bundlers) build.
+  {
+    input: 'src/main.js',
+    output: [
+      { file: pkg.main, format: 'cjs' },
+      { file: pkg.module, format: 'es' },
+    ],
+    external: ['lodash.isplainobject'],
+    plugins: [
+      babel({
+        exclude: ['node_modules/**'],
+      }),
+    ],
+  },
+]