+++ packages/util-nonempty/changelog.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.4.0] - 2017-09-23
+### Added
+- ✨ Implemented Rollup to generate 3 flavours of this package: CommonJS, UMD and ESM `module` with `import`/`export`.
+- ✨ [Automatic](https://github.com/camacho/format-package) `package.json` sorting.
+
 ## [1.3.0] - 2017-09-18
 ### Added
 - ✨ This changelog.
@@ -13,4 +18,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### New
 - ✨ First public release
 
+[1.4.0]: https://github.com/codsen/util-nonempty/compare/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/codsen/util-nonempty/compare/v1.3.0...v1.3.0